fixed chicken not healing player
first division was rounded down to 0 due to being an int -> changed to float then cast back to int

+++ Assets/Scripts/Enemies/ChickenCombatHandler.cs
@@ -36,6 +36,6 @@ public class ChickenCombatHandler : EnemyCombatHandler
     public void RewardPlayer()
     {
         print("Rewarding boss kill");
-        FindObjectOfType<PlayerStats>().CurrentHealth += (FindObjectOfType<PlayerStats>().MaxHealth / 100) * percentHealthRewarded; //issue #54
+        FindObjectOfType<PlayerStats>().CurrentHealth += (int)(FindObjectOfType<PlayerStats>().MaxHealth / 100f * percentHealthRewarded); //issue #54
     }
 }
